Move data download dialog to shared UI module
