+++ ui/projects/streampipes/shared-ui/src/lib/components/configuration-box/configuration-box.component.ts
