+++ ui/projects/streampipes/shared-ui/src/lib/components/date-input/date-input.component.ts
