+++ ui/projects/streampipes/shared-ui/src/lib/dialog/data-download-dialog/components/download/download.component.html
