--- ui/src/app/core-ui/data-download-dialog/components/select-format/select-format.component.html
@@ -1,127 +1,0 @@
-<!--
-  ~  Licensed to the Apache Software Foundation (ASF) under one or more
-  ~  contributor license agreements.  See the NOTICE file distributed with
-  ~  this work for additional information regarding copyright ownership.
-  ~  The ASF licenses this file to You under the Apache License, Version 2.0
-  ~  (the "License"); you may not use this file except in compliance with
-  ~  the License.  You may obtain a copy of the License at
-  ~
-  ~     http://www.apache.org/licenses/LICENSE-2.0
-  ~
-  ~  Unless required by applicable law or agreed to in writing, software
-  ~  distributed under the License is distributed on an "AS IS" BASIS,
-  ~  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  ~  See the License for the specific language governing permissions and
-  ~  limitations under the License.
-  ~
-  -->
-
-<ng-template matStepLabel>Select Format</ng-template>
-<div>
-    <h5>Download Format</h5>
-    <mat-radio-group
-        class="sp-radio-group"
-        [(ngModel)]="formatExportConfig.format"
-    >
-        <mat-radio-button
-            value="json"
-            class="sp-radio-button"
-            data-cy="download-configuration-json"
-        >
-            JSON
-        </mat-radio-button>
-        <mat-radio-button
-            value="csv"
-            class="sp-radio-button"
-            data-cy="download-configuration-csv"
-        >
-            CSV
-        </mat-radio-button>
-        <mat-radio-button
-            value="xlsx"
-            class="sp-radio-button"
-            data-cy="download-configuration-xlsx"
-        >
-            Excel
-        </mat-radio-button>
-    </mat-radio-group>
-</div>
-<div *ngIf="formatExportConfig.format === 'csv'" class="mt-10">
-    <h5>Delimiter</h5>
-    <mat-radio-group
-        [(ngModel)]="formatExportConfig.delimiter"
-        class="sp-radio-group"
-    >
-        <mat-radio-button
-            value="comma"
-            class="sp-radio-button"
-            data-cy="download-configuration-delimiter-comma"
-            >&nbsp;,
-        </mat-radio-button>
-        <mat-radio-button
-            value="semicolon"
-            class="sp-radio-button"
-            data-cy="download-configuration-delimiter-semicolon"
-            >&nbsp;;
-        </mat-radio-button>
-    </mat-radio-group>
-</div>
-<div
-    *ngIf="formatExportConfig.format === 'xlsx' && hasReadFilePrivilege"
-    fxLayout="column"
->
-    <mat-checkbox
-        [(ngModel)]="formatExportConfig.useTemplate"
-        [disabled]="excelTemplates.length === 0"
-    >
-        Use uploaded file template
-    </mat-checkbox>
-    @if (formatExportConfig.useTemplate && excelTemplates.length > 0) {
-        <mat-form-field class="mt-10" color="accent">
-            <mat-select
-                [(ngModel)]="formatExportConfig.templateId"
-                placeholder="Choose template"
-            >
-                <mat-option
-                    *ngFor="let template of excelTemplates"
-                    [value]="template.fileId"
-                >
-                    {{ template.filename }}
-                </mat-option>
-            </mat-select>
-        </mat-form-field>
-        <mat-form-field color="accent">
-            <mat-label>First row index to append data</mat-label>
-            <input
-                matInput
-                [(ngModel)]="formatExportConfig.startRow"
-                type="number"
-            />
-        </mat-form-field>
-    }
-</div>
-<div
-    *ngIf="
-        formatExportConfig.format === 'xlsx' ||
-        formatExportConfig.format === 'csv'
-    "
->
-    <h5>Header column name</h5>
-    <mat-radio-group
-        [(ngModel)]="formatExportConfig.headerColumnName"
-        class="sp-radio-group"
-    >
-        <mat-radio-button
-            value="key"
-            class="sp-radio-button"
-            data-cy="download-configuration-column-name-key"
-            >Use field key (runtime name) as header column
-        </mat-radio-button>
-        <mat-radio-button
-            value="label"
-            class="sp-radio-button"
-            data-cy="download-configuration-column-name-label"
-            >Use field label as header column if available
-        </mat-radio-button>
-    </mat-radio-group>
-</div>
